:refactor: Simplificar a configuração de autenticação, movendo a lógica do NextAuth para um arquivo separado e ajustando as importações nos arquivos de rotas. Atualizar a middleware para excluir a rota de registro, garantindo uma configuração mais limpa e eficiente.
