+++ app/api/projects/active/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
 
 import { db } from "@/lib/db";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 
 // PUT /api/projects/active - Definir o projeto ativo do usuário
 export async function PUT(req: NextRequest) {
