+++ app/api/projects/members/[...params]/route.ts
@@ -2,7 +2,7 @@ import { NextResponse, NextRequest } from "next/server";
 import { getServerSession } from "next-auth";
 
 import { db } from "@/lib/db";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 
 type RoleType = "owner" | "admin" | "member" | "viewer";
 type RolePriorityMap = {
